+++ examples/exmp008_relativ_corr/job.py
@@ -38,6 +38,7 @@ def run_exmp008(
     output = calc.get_output()
     if not output.terminated_normally():
         print(f"ORCA calculation failed, see output file: {output.get_outfile()}")
+        print(output.error_message())
         sys.exit(1)
     # << END OF IF