+++ examples/exmp010_uvvis/job.py
@@ -36,6 +36,7 @@ def run_exmp010(
     output = calc.get_output()
     if not output.terminated_normally():
         print(f"ORCA calculation failed, see output file: {output.get_outfile()}")
+        print(output.error_message())
         sys.exit(1)
     # << END OF IF