+++ examples/exmp011_epr/job.py
@@ -45,6 +45,7 @@ def run_exmp011(
     output = calc.get_output()
     if not output.terminated_normally():
         print(f"ORCA calculation failed, see output file: {output.get_outfile()}")
+        print(output.error_message())
         sys.exit(1)
     # << END OF IF