+++ examples/exmp013_bs/job.py
@@ -34,6 +34,7 @@ def run_exmp013(
     output = calc.get_output()
     if not output.terminated_normally():
         print(f"ORCA calculation failed, see output file: {output.get_outfile()}")
+        print(output.error_message())
         sys.exit(1)
     # << END OF IF