+++ examples/exmp021_basis/job.py
@@ -37,6 +37,7 @@ def run_exmp021(
     output = calc.get_output()
     if not output.terminated_normally():
         print(f"ORCA calculation failed, see output file: {output.get_outfile()}")
+        print(output.error_message())
         sys.exit(1)
     # << END OF IF