+++ examples/exmp028_nevpt2/job.py
@@ -38,8 +38,12 @@ def run_exmp028(
     output = calc.get_output()
     if not output.terminated_normally():
         print(f"ORCA calculation failed, see output file: {output.get_outfile()}")
+        print(output.error_message())
         sys.exit(1)
-    # << END OF IF
+    if not output.casscf_converged():
+        print(f"ORCA calculation failed, see output file: {output.get_outfile()}")
+        print(output.error_message())
+        sys.exit(1)
 
     # > Parse JSON files
     output.parse()