+++ examples/exmp036_solvator/job.py
@@ -37,6 +37,7 @@ def run_exmp036(
     output = calc.get_output()
     if not output.terminated_normally():
         print(f"ORCA calculation failed, see output file: {output.get_outfile()}")
+        print(output.error_message())
         sys.exit(1)
 
     output.parse(read_gbw_json=False)