+++ examples/exmp044_from_ase/job.py
@@ -49,6 +49,7 @@ def run_exmp044(working_dir: Path | None = Path("RUN")) -> Output:
     output = calc.get_output()
     if not output.terminated_normally():
         print(f"ORCA calculation failed, see output file: {output.get_outfile()}")
+        print(output.error_message())
         sys.exit(1)
     # << END OF IF