+++ examples/exmp045_existing_calc/job.py
@@ -17,6 +17,7 @@ def run_exmp045() -> Output:
     output = calc.get_output()
     if not output.terminated_normally():
         print(f"ORCA calculation failed, see output file: {output.get_outfile()}")
+        print(output.error_message())
         sys.exit(1)
     # << END OF IF