+++ examples/exmp047_trj/job.py
@@ -41,6 +41,7 @@ def run_exmp047(working_dir: Path | None = Path("RUN")) -> list[Output]:
         output = calc.get_output()
         if not output.terminated_normally():
             print(f"ORCA calculation failed, see output file: {output.get_outfile()}")
+            print(output.error_message())
             sys.exit(1)
         # << END OF IF