+++ examples/exmp053_opencosmors/job.py
@@ -32,6 +32,7 @@ def run_exmp053(
     output = calc.get_output()
     if not output.terminated_normally():
         print(f"ORCA calculation failed, see output file: {output.get_outfile()}")
+        print(output.error_message())
         sys.exit(1)
     # << END OF IF