+++ src/opi/output/core.py
@@ -18,8 +18,12 @@ from opi.input.structures import Atom, Coordinates, Structure
 from opi.output.cube import CubeOutput
 from opi.output.gbw_suffix import GbwSuffix
 from opi.output.grepper.recipes import (
+    get_error_message,
+    get_error_messages,
     get_float_from_line,
     get_lines_from_block,
+    has_casscf_converged,
+    has_cc_converged,
     has_geometry_optimization_converged,
     has_scf_converged,
     has_terminated_normally,
@@ -666,11 +670,47 @@ class Output:
         bool
             True if "ORCA TERMINATED NORMALLY" is found in ".out" file else False
         """
+        return self._has(has_terminated_normally)
+
+    def _has(self, has_func: Callable[[Path], bool]) -> bool:
         outfile = self.get_outfile()
         try:
-            return has_terminated_normally(outfile)
+            return has_func(outfile)
         except FileNotFoundError:
             return False
+
+    def error_messages(self) -> list[str]:
+        """
+        Return all known error messages found in the ORCA output file.
+        Scanning stops early when a critical pattern is matched. Critical patterns are patterns that directly terminate ORCA.
+        If the output file does not exist a corresponding error message is returned.
+
+        Returns
+        -------
+        list[str] | None
+            List of error message strings if any errors were detected, else None.
+        """
+        outfile = self.get_outfile()
+        try:
+            return get_error_messages(outfile)
+        except FileNotFoundError:
+            return [f"Could not find output file: {outfile}"]
+
+    def error_message(self) -> str:
+        """
+        Return the most important error message found in the ORCA output file or an empty string if no known error was detected.
+        If the output file does not exist a corresponding error message is returned.
+
+        Returns
+        -------
+        str | None
+            Error message string if an error was detected, else None.
+        """
+        outfile = self.get_outfile()
+        try:
+            return get_error_message(outfile)
+        except FileNotFoundError:
+            return f"Could not find output file: {outfile}"
 
     def scf_converged(self) -> bool:
         """
@@ -683,11 +723,33 @@ class Output:
         bool
             True if "SUCCESS" is found in ".out" file else False
         """
-        outfile = self.get_outfile()
-        try:
-            return has_scf_converged(outfile)
-        except FileNotFoundError:
-            return False
+        return self._has(has_scf_converged)
+
+    def casscf_converged(self) -> bool:
+        """
+        Determine if ORCA CAS-SCF converged, by looking for "THE CAS-SCF GRADIENT HAS CONVERGED" in the ".out" file.
+        Check only if ORCA CAS-SCF was actually requested.
+        If the ".out" file does not exist, also return False.
+
+        Returns
+        -------
+        bool
+            True if string is found in ".out" file else False
+        """
+        return self._has(has_casscf_converged)
+
+    def cc_converged(self) -> bool:
+        """
+        Determine if ORCA coupled-cluster iterations converged, by looking for "The Coupled-Cluster iterations have converged" in the ".out" file.
+        Check only if CC was actually requested.
+        If the ".out" file does not exist, also return False.
+
+        Returns
+        -------
+        bool
+            True if string is found in ".out" file else False
+        """
+        return self._has(has_cc_converged)
 
     def geometry_optimization_converged(self) -> bool:
         """
@@ -700,11 +762,7 @@ class Output:
         bool
             True if "HURRAY" is found in ".out" file else False
         """
-        outfile = self.get_outfile()
-        try:
-            return has_geometry_optimization_converged(outfile)
-        except FileNotFoundError:
-            return False
+        return self._has(has_geometry_optimization_converged)
 
     def print_graph(self, *, max_length: int = 3, depth: int = -1) -> None:
         """